[feat] 프로필 사진 변경 기능 추가(#105)
Close #105

+++ src/main/java/hyangyu/server/aws/S3Uploader.java
@@ -1,6 +1,7 @@
 package hyangyu.server.aws;
 
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
@@ -33,6 +34,8 @@ public class S3Uploader {
     }
     
     public void delete(String name) {
+    	if(!amazonS3Client.doesObjectExist(bucket, name))
+    		 throw new AmazonS3Exception("Object " +name+ " does not exist!");
     	amazonS3Client.deleteObject(bucket,name);
     }
 
@@ -52,13 +55,6 @@ public class S3Uploader {
     
     public String getThumbnailPath(String path) {
         return amazonS3Client.getUrl(bucket, path).toString();
-    }
-    
-    public String findImg(String name) {
-    	if(amazonS3Client.getObject(bucket, name) != null)
-    		return amazonS3Client.getObject(bucket, name).toString();
-    	else
-    		return null;
     }
 
     // 로컬에 저장된 이미지 지우기
